+++ src/Kanbas/Courses/Modules/index.tsx
@@ -3,8 +3,7 @@ import ModuleControlButtons from "./ModuleControlButtons";
 import ModulesControls from "./ModulesControls";
 import { BsGripVertical } from "react-icons/bs";
 import { useParams } from "react-router";
-import * as db from "../../Database"; // Default import
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import {
   setModules,
   addModule,
@@ -15,43 +14,68 @@ import {
 import { useSelector, useDispatch } from "react-redux";
 import * as coursesClient from "../client";
 import * as modulesClient from "./client";
+import { RootState } from "../../store";
+import { Module } from "../../types";
 
 export default function Modules() {
-  const { cid } = useParams<{ cid: string }>(); // Ensure the type is specified for useParams
+  const { cid } = useParams<{ cid: string }>();
   const [moduleName, setModuleName] = useState("");
-  const { modules } = useSelector((state: any) => state.modulesReducer);
+  const { modules } = useSelector((state: RootState) => state.modulesReducer);
   const dispatch = useDispatch();
+  const [error, setError] = useState("");
+
   const fetchModules = async () => {
-    const modules = await coursesClient.findModulesForCourse(cid as string);
-    dispatch(setModules(modules));
+    if (!cid) return;
+    try {
+      const modules = await coursesClient.findModulesForCourse(cid);
+      dispatch(setModules(modules));
+    } catch (err) {
+      console.error(err);
+      setError("Unable to load modules. Please try again.");
+    }
   };
   useEffect(() => {
     fetchModules();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
   const createModuleForCourse = async () => {
     if (!cid) return;
-    const newModule = { name: moduleName, course: cid };
-    const module = await coursesClient.createModuleForCourse(cid, newModule);
-    dispatch(addModule(module));
+    try {
+      const newModule = { name: moduleName, course: cid };
+      const module = await coursesClient.createModuleForCourse(cid, newModule);
+      dispatch(addModule(module));
+      setModuleName("");
+    } catch (err) {
+      console.error(err);
+      setError("Unable to create module. Please try again.");
+    }
   };
   const removeModule = async (moduleId: string) => {
-    await modulesClient.deleteModule(moduleId);
-    dispatch(deleteModule(moduleId));
+    if (!window.confirm("Are you sure you want to delete this module?")) {
+      return;
+    }
+    try {
+      await modulesClient.deleteModule(moduleId);
+      dispatch(deleteModule(moduleId));
+    } catch (err) {
+      console.error(err);
+      setError("Unable to delete module. Please try again.");
+    }
   };
-  const saveModule = async (module: any) => {
-    await modulesClient.updateModule(module);
-    dispatch(updateModule(module));
+  const saveModule = async (module: Module) => {
+    try {
+      await modulesClient.updateModule(module);
+      dispatch(updateModule(module));
+    } catch (err) {
+      console.error(err);
+      setError("Unable to save module. Please try again.");
+    }
   };
 
   return (
     <div className="ms-4">
-      {/* <ModulesControls
-        setModuleName={setModuleName}
-        moduleName={moduleName}
-        addModule={addModule}
-        moduleId={module._id}
-        deleteModule={deleteModule}
-      /> */}
+      {error && <div className="alert alert-danger">{error}</div>}
       <ModulesControls
         moduleName={moduleName}
         setModuleName={setModuleName}
@@ -62,53 +86,51 @@ export default function Modules() {
       <br />
       <br />
       <ul id="wd-modules" className="list-group rounded-0">
-        {modules
-          // .filter((module: any) => module.course === cid)
-          .map((module: any) => (
-            <li
-              key={module.id}
-              className="wd-module list-group-item p-0 mb-5 fs-5 border-gray"
-            >
-              <div className="wd-title p-3 ps-2 bg-secondary">
-                <BsGripVertical className="me-2 fs-3" />{" "}
-                {!module.editing && module.name}
-                {module.editing && (
-                  <input
-                    className="form-control w-50 d-inline-block"
-                    onChange={(e) =>
-                      dispatch(
-                        updateModule({ ...module, name: e.target.value })
-                      )
+        {modules.map((module) => (
+          <li
+            key={module._id}
+            className="wd-module list-group-item p-0 mb-5 fs-5 border-gray"
+          >
+            <div className="wd-title p-3 ps-2 bg-secondary">
+              <BsGripVertical className="me-2 fs-3" />{" "}
+              {!module.editing && module.name}
+              {module.editing && (
+                <input
+                  className="form-control w-50 d-inline-block"
+                  onChange={(e) =>
+                    dispatch(
+                      updateModule({ ...module, name: e.target.value })
+                    )
+                  }
+                  onKeyDown={(e) => {
+                    if (e.key === "Enter") {
+                      saveModule({ ...module, editing: false });
                     }
-                    onKeyDown={(e) => {
-                      if (e.key === "Enter") {
-                        saveModule({ ...module, editing: false });
-                      }
-                    }}
-                    defaultValue={module.name}
-                  />
-                )}
-                <ModuleControlButtons
-                  moduleId={module._id}
-                  deleteModule={(moduleId) => removeModule(moduleId)}
-                  editModule={(moduleId) => dispatch(editModule(moduleId))}
+                  }}
+                  defaultValue={module.name}
                 />
-              </div>
-              {module.lessons && (
-                <ul className="wd-lessons list-group rounded-0">
-                  {module.lessons.map((lesson: any) => (
-                    <li
-                      key={lesson.id}
-                      className="wd-lesson list-group-item p-3 ps-1"
-                    >
-                      <BsGripVertical className="me-2 fs-3" /> {lesson.name}{" "}
-                      <LessonControlButtons />
-                    </li>
-                  ))}
-                </ul>
               )}
-            </li>
-          ))}
+              <ModuleControlButtons
+                moduleId={module._id}
+                deleteModule={(moduleId) => removeModule(moduleId)}
+                editModule={(moduleId) => dispatch(editModule(moduleId))}
+              />
+            </div>
+            {module.lessons && (
+              <ul className="wd-lessons list-group rounded-0">
+                {module.lessons.map((lesson) => (
+                  <li
+                    key={lesson._id}
+                    className="wd-lesson list-group-item p-3 ps-1"
+                  >
+                    <BsGripVertical className="me-2 fs-3" /> {lesson.name}{" "}
+                    <LessonControlButtons />
+                  </li>
+                ))}
+              </ul>
+            )}
+          </li>
+        ))}
       </ul>
     </div>
   );
